更新: apm 添加平台切换
